+++ finalize.sh
@@ -1,0 +1,49 @@
+#!/bin/bash
+#
+# Finalize the MassOS build. Stage 3 will run this in chroot once the Stage 3
+# build is finished.
+set -e
+# Ensure we're running in the MassOS chroot.
+if [ $EUID -ne 0 ] || [ ! -d /sources ]; then
+  echo "DO NOT RUN THIS SCRIPT ON YOUR HOST SYSTEM." >&2
+  echo "IT WILL RENDER YOUR SYSTEM UNUSABLE." >&2
+  echo "YOU HAVE BEEN WARNED!!!" >&2
+  exit 1
+fi
+# Uninstall Rust.
+if [ -x /usr/lib/rustlib/uninstall.sh ]; then
+  /usr/lib/rustlib/uninstall.sh
+fi
+# Compress manual pages.
+zman /usr/share/man
+# Remove leftover junk in /root.
+rm -rf /root/.{cache,cargo,cmake}
+# Remove Debian stuff.
+rm -rf /etc/kernel
+# Move any misplaced files.
+if [ -d /usr/etc ]; then
+  cp -r /usr/etc /
+  rm -rf /usr/etc
+fi
+if [ -d /usr/man ]; then
+  cp -r /usr/man /usr/share
+  rm -rf /usr/man
+fi
+# Remove static documentation to free up space.
+rm -rf /usr/share/doc/*
+rm -rf /usr/doc
+rm -rf /usr/docs
+rm -rf /usr/share/gtk-doc/html/*
+# Remove temporary compiler from stage1.
+find /usr -depth -name $(uname -m)-massos-linux-gnu\* | xargs rm -rf
+# Remove libtool archives.
+find /usr/lib /usr/libexec -name \*.la -delete
+# Remove any temporary files.
+rm -rf /tmp/*
+# As a finishing touch, run ldconfig and other misc commands.
+ldconfig
+glib-compile-schemas /usr/share/glib-2.0/schemas
+gtk-update-icon-cache -q -t -f --include-image-data /usr/share/icons/hicolor
+update-desktop-database
+# Clean up and self-destruct.
+rm -rf /sources