+++ stage2-resume-failed.sh
@@ -26,25 +26,8 @@ if [ ! -d "$MASSOS" ]; then
 fi
 # Chroot into the MassOS environment and continue the build.
 utils/programs/mass-chroot "$MASSOS" /sources/build-system.sh
-# Strip executables and libraries to free up space.
-printf "Stripping binaries... "
-find "$MASSOS"/usr/{bin,libexec,sbin} -type f -exec strip --strip-all {} ';' &> /dev/null || true
-echo "Done!"
-printf "Stripping libraries... "
-find "$MASSOS"/usr/lib -type f -name \*.a -exec strip --strip-debug {} ';' &> /dev/null || true
-find "$MASSOS"/usr/lib -type f -name \*.so\* -exec strip --strip-unneeded {} ';' &> /dev/null || true
-echo "Done!"
-# Finish the MassOS system.
-outfile="massos-$(cat utils/massos-release)-rootfs-x86_64.tar"
-printf "Creating $outfile... "
-cd "$MASSOS"
-tar -cpf ../"$outfile" *
-cd ..
-echo "Done!"
-echo "Compressing $outfile with XZ (using $(nproc) threads)..."
-xz -v --threads=$(nproc) "$outfile"
-echo "Successfully created $outfile.xz."
-printf "SHA256 checksum: "
-sha256sum $outfile.xz | sed "s/  $outfile.xz//"
-# Clean up.
-rm -rf $MASSOS
+# Finishing message.
+echo
+echo "Stage 2 build completed successfully."
+echo "You must now run stage3.sh and pass a supported desktop environment as"
+echo "an argument. See 'stage3/README' for more information."