+++ stage3/nodesktop/stage3-nodesktop.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+# Just stall for a few seconds so the chroot doesn't bug out.
+sleep 3